feat: implement login protection (#282)
* feat: implement login protection

adds the enable_login_protection field to the api ui page and
/api/v1/system/api. this field toggles enabling login protection
for api authentication. this will help prevent brute force
attacks against api endpoints

* fix: check if enable_login_protection is empty

on the api ui page, we were evaluating a base variable which
erroneously set or unset the variable

* fix: default auth logs username

login protection requires the username to be set in order to work.
this sets the default username to 'unknown' if no username could
be found

* chore: enable login protection by default

* tests: added login protection test

* lint: fixed import order for tests

* tests: just expect login protection timeout

before the test_login_protection test expected requests to start failing at a
specific point, but this does not work with all login protection configs and
situations. it now just expects the requests to start to timeout after many
failed auth attempts