chore(privs): add unique priv for service_watchdog
currently pfsense does not include a unique
privilege for the service watchdog package.
this forces clients to hold the page-all
privilege to ever interact with the endpoint
which poses a security issue. this commit
adds a unique privilege for the service watchdog
api endpoint. this privilege does not allow
users access to the service watchdog ui page.