perf: only perform query if query params were given
This ensures queries are only performed if the client provided query parameters. Queries are more resource intensive since we have to load all objects to perform the query, even if pagination is used. If only pagination is needed and no query parameters were given, we now delegate the job to read_all(), which allows only the paginated objects to be loaded.