fix: skip query fields that don't exist
Before if you queried a field that did not exist for
the given Model, it would prevent that Model from
matching the query criteria. However, there may be
situation where a query is intended for multiple
types of Models that may not have the same fields.
In this situation, we should allow the query to
continue.