fix: clear gateway cache before applying routing changes
pfSense 2.8.0 adds a new backend GatewaysCache that is not automatically kept up-to-date. This means there can be a conflict between the current configuration and the cache when the routing subsystem is applied which can result in duplicate or unintentional routing changes. This change clears the cache before we apply routing changes to ensure its always up-to-date with the config